requirements.txt
fastapi
sqlalchemy
psycopg2-binary
uvicorn
sqlmodel
